# perf: Remove atomic